+++ MTManger/composeApp/src/commonMain/kotlin/com/wonddak/mtmanger/ui/view/home/person/PersonView.kt
@@ -1,4 +1,4 @@
-package com.wonddak.mtmanger.ui.view.home
+package com.wonddak.mtmanger.ui.view.home.person
 
 import androidx.compose.foundation.BorderStroke
 import androidx.compose.foundation.ExperimentalFoundationApi
@@ -19,6 +19,8 @@ import androidx.compose.foundation.shape.RoundedCornerShape
 import androidx.compose.material3.Card
 import androidx.compose.material3.CardDefaults
 import androidx.compose.material3.HorizontalDivider
+import androidx.compose.material3.Icon
+import androidx.compose.material3.IconButton
 import androidx.compose.material3.OutlinedButton
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
@@ -30,30 +32,31 @@ import androidx.compose.runtime.setValue
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.draw.clip
-import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.text.style.TextAlign
-import androidx.compose.ui.unit.TextUnit
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
 import com.wonddak.mtmanger.model.Resource
-import com.wonddak.mtmanger.room.entity.BuyGood
-import com.wonddak.mtmanger.room.entity.MtDataList
+import com.wonddak.mtmanger.room.entity.*
+import com.wonddak.mtmanger.toPriceString
 import com.wonddak.mtmanger.ui.theme.maple
 import com.wonddak.mtmanger.ui.theme.match1
 import com.wonddak.mtmanger.ui.theme.match2
-import com.wonddak.mtmanger.ui.view.common.DefaultText
 import com.wonddak.mtmanger.ui.view.common.FeeInfo
-import com.wonddak.mtmanger.ui.view.dialog.BuyDialog
 import com.wonddak.mtmanger.ui.view.dialog.DeleteDialog
+import com.wonddak.mtmanger.ui.view.dialog.PersonDialog
+import com.wonddak.mtmanger.ui.view.home.buy.BuyGoodItemText
+import com.wonddak.mtmanger.util.DeviceActionHelper
 import com.wonddak.mtmanger.viewModel.MTViewModel
 import mtmanger.composeapp.generated.resources.Res
 import mtmanger.composeapp.generated.resources.dialog_delete_reset
+import mtmanger.composeapp.generated.resources.ic_baseline_phone_24
 import mtmanger.composeapp.generated.resources.info_text
+import org.jetbrains.compose.resources.painterResource
 import org.jetbrains.compose.resources.stringResource
-
+import org.koin.compose.koinInject
 
 @Composable
-fun BuyView(
+fun PersonView(
     mtViewModel: MTViewModel,
 ) {
     Column(
@@ -71,7 +74,7 @@ fun BuyView(
                     .padding(vertical = 5.dp)
             )
             Text(
-                text = "구매내역",
+                text = "참여자 명단",
                 color = match2,
                 fontSize = 30.sp,
                 fontFamily = maple(),
@@ -92,21 +95,21 @@ fun BuyView(
                 color = match2
             )
             Spacer(modifier = Modifier.height(3.dp))
-            BuyItemList(
-                Modifier.fillMaxHeight(),
-                mtViewModel
+            PersonItemList(
+                Modifier.fillMaxHeight(), mtViewModel
             )
         }
         Row {
-            BuyGoodPanel(mtViewModel)
+            PersonPanel(mtViewModel)
         }
     }
 }
 
 @Composable
-fun BuyGoodPanel(
+fun PersonPanel(
     mtViewModel: MTViewModel,
 ) {
+//    val context = LocalContext.current
     val resource: Resource<MtDataList> by mtViewModel.nowMtDataList.collectAsState()
     var showItemReset by remember {
         mutableStateOf(false)
@@ -116,6 +119,8 @@ fun BuyGoodPanel(
     }
     if (resource is Resource.Success) {
         (resource as Resource.Success<MtDataList>).data?.let { mtData ->
+
+
             Column(
                 modifier = Modifier
                     .fillMaxWidth()
@@ -141,37 +146,28 @@ fun BuyGoodPanel(
                         horizontalAlignment = Alignment.CenterHorizontally,
                         verticalArrangement = Arrangement.SpaceAround
                     ) {
-                        val sumOfGoodsFee = mtData.buyGoodList.sumOf { it.price * it.count }
-                        val sumOfPersonPayFee = mtData.personList.sumOf { it.paymentFee }
                         val modifier = Modifier.padding(3.dp)
                         FeeInfo(
-                            modifier,
-                            text = "지출 금액",
-                            fee = sumOfGoodsFee
+                            modifier, text = "총 참여자", fee = mtData.personList.size, feeIndex = "명"
                         )
-                        FeeInfo(
-                            modifier,
-                            text = "남은 금액",
-                            fee = sumOfPersonPayFee - sumOfGoodsFee
-                        )
+                        FeeInfo(modifier,
+                            text = "받은 금액",
+                            fee = mtData.personList.sumOf { it.paymentFee })
                     }
                     Column(
-                        Modifier.weight(2f),
-                        horizontalAlignment = Alignment.CenterHorizontally
+                        Modifier.weight(2f), horizontalAlignment = Alignment.CenterHorizontally
                     ) {
                         OutlinedButton(
                             modifier = Modifier.fillMaxWidth(),
                             onClick = { showAddDialog = true },
                             border = BorderStroke(2.dp, match2)
                         ) {
-                            BuyGoodItemText(text = "내역 추가")
+                            BuyGoodItemText(text = "참여자 추가")
                         }
                         OutlinedButton(
-                            modifier = Modifier.fillMaxWidth(),
-                            onClick = {
+                            modifier = Modifier.fillMaxWidth(), onClick = {
                                 showItemReset = true
-                            },
-                            border = BorderStroke(2.dp, match2)
+                            }, border = BorderStroke(2.dp, match2)
                         ) {
                             BuyGoodItemText(text = "초기화")
                         }
@@ -185,7 +181,7 @@ fun BuyGoodPanel(
         DeleteDialog(
             msg = stringResource(resource = Res.string.dialog_delete_reset),
             onDelete = {
-                mtViewModel.clearBuyGoodData()
+                mtViewModel.clearPersonData()
                 showItemReset = false
             },
             onDismiss = {
@@ -193,16 +189,14 @@ fun BuyGoodPanel(
             },
         )
     }
-    val categoryList by mtViewModel.settingCategoryList.collectAsState(emptyList())
-
     if (showAddDialog) {
-        BuyDialog(
-            categoryList = categoryList,
+        PersonDialog(
+            null,
             onDismiss = {
                 showAddDialog = false
             },
             onAdd = { data ->
-                mtViewModel.insertBuyGood(data)
+                mtViewModel.insertPerson(data)
                 showAddDialog = false
             }
         )
@@ -210,15 +204,13 @@ fun BuyGoodPanel(
 }
 
 @Composable
-fun BuyItemList(
+fun PersonItemList(
     modifier: Modifier = Modifier,
     mtViewModel: MTViewModel,
 ) {
     val resource: Resource<MtDataList> by mtViewModel.nowMtDataList.collectAsState()
     Card(
-        modifier = modifier,
-        shape = RoundedCornerShape(10.dp),
-        colors = CardDefaults.cardColors(
+        modifier = modifier, shape = RoundedCornerShape(10.dp), colors = CardDefaults.cardColors(
             containerColor = match2
         )
     ) {
@@ -233,19 +225,16 @@ fun BuyItemList(
                     .padding(vertical = 5.dp)
 
             ) {
-                listOf(
-                    "분류",
-                    "이름",
-                    "수량",
-                    "단가",
-                    "합"
-                ).forEach {title ->
-                    BuyGoodItemText(
-                        Modifier.weight(1f),
-                        color = match1,
-                        text = title
-                    )
-                }
+                val weight1 = Modifier.weight(1f)
+                BuyGoodItemText(
+                    weight1, color = match1, text = "이름"
+                )
+                BuyGoodItemText(
+                    weight1, color = match1, text = "납부금액"
+                )
+                BuyGoodItemText(
+                    weight1, color = match1, text = "폰"
+                )
             }
             Card(
                 modifier = Modifier
@@ -255,19 +244,18 @@ fun BuyItemList(
                 colors = CardDefaults.cardColors(
                     containerColor = match1
                 )
-            )
-            {
+            ) {
                 if (resource is Resource.Success) {
                     (resource as Resource.Success<MtDataList>).data?.let {
-                        val buyGoodList = it.buyGoodList
+                        val personList = it.personList
                         LazyColumn(
                             modifier = Modifier
                                 .fillMaxWidth()
                                 .padding(vertical = 5.dp)
                         ) {
-                            itemsIndexed(buyGoodList) { index, buyGood ->
-                                BuyItemView(buyGood, mtViewModel)
-                                if (index != buyGoodList.size - 1) {
+                            itemsIndexed(personList) { index, person ->
+                                PersonItemView(person, mtViewModel)
+                                if (index != personList.size - 1) {
                                     HorizontalDivider(
                                         color = match2
                                     )
@@ -283,8 +271,8 @@ fun BuyItemList(
 
 @OptIn(ExperimentalFoundationApi::class)
 @Composable
-fun BuyItemView(
-    buyGood: BuyGood,
+fun PersonItemView(
+    person: Person,
     mtViewModel: MTViewModel,
 ) {
     var showItemDelete by remember {
@@ -293,6 +281,7 @@ fun BuyItemView(
     var showEditDialog by remember {
         mutableStateOf(false)
     }
+    val deviceActionHelper : DeviceActionHelper = koinInject()
     Row(
         modifier = Modifier
             .fillMaxWidth()
@@ -304,13 +293,25 @@ fun BuyItemView(
                 onLongClick = {
                     showItemDelete = true
                 },
-            ),
-        verticalAlignment = Alignment.CenterVertically
+            ), verticalAlignment = Alignment.CenterVertically
     ) {
-        buyGood.getItemList().forEach { title ->
-            BuyGoodItemText(
-                Modifier.weight(1f),
-                text = title
+        val weight1 = Modifier.weight(1f)
+        BuyGoodItemText(
+            weight1, text = person.name
+        )
+        BuyGoodItemText(
+            weight1, text = person.paymentFee.toPriceString()
+        )
+        IconButton(
+            modifier = weight1,
+            onClick = {
+                deviceActionHelper.makeCall(person.phoneNumber)
+            }
+        ) {
+            Icon(
+                painter = painterResource(resource = Res.drawable.ic_baseline_phone_24),
+                contentDescription = null,
+                tint = match2
             )
         }
     }
@@ -318,7 +319,7 @@ fun BuyItemView(
     if (showItemDelete) {
         DeleteDialog(
             onDelete = {
-                mtViewModel.deleteBuyGood(buyGood.buyGoodId!!)
+                mtViewModel.deletePerson(person.personId!!)
                 showItemDelete = false
             },
             onDismiss = {
@@ -327,35 +328,15 @@ fun BuyItemView(
         )
     }
     if (showEditDialog) {
-        val categoryList by mtViewModel.settingCategoryList.collectAsState(emptyList())
-        BuyDialog(
-            buyGood = buyGood,
-            categoryList = categoryList,
+        PersonDialog(
+            person,
             onDismiss = {
                 showEditDialog = false
             },
             onAdd = { data ->
-                mtViewModel.insertBuyGood(
-                    data,
-                    buyGood.buyGoodId!!
-                )
+                mtViewModel.updatePerson(person.personId!!, data)
                 showEditDialog = false
             }
         )
     }
-}
-
-@Composable
-fun BuyGoodItemText(
-    modifier: Modifier = Modifier,
-    color: Color = match2,
-    text: String,
-    fontSize: TextUnit = 12.sp,
-) {
-    DefaultText(
-        modifier = modifier,
-        color = color,
-        text = text,
-        fontSize = fontSize,
-    )
 }
